requirements.txt
PyQt5==5.15.10
chromadb==0.4.15
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3
google-generativeai==0.8.5
huggingface-hub>=0.17.3
psutil==5.9.6
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
textblob==0.17.1
openpyxl==3.1.2
python-pptx==0.6.21
reportlab==4.0.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl